% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{aggregateMDT}
\alias{aggregateMDT}
\alias{aggregateMDT,MDT-method}
\title{Aggregate FeatureIDs Into Higher Level IDs}
\usage{
aggregateMDT(x, group, fun.aggregate = function(value) mean(value, na.rm =
  TRUE), ...)

\S4method{aggregateMDT}{MDT}(x, group, fun.aggregate = function(value)
  mean(value, na.rm = TRUE), ...)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{group}{\code{data.frame} specifying how to aggregate \code{FeatureID}s.
The first column is the old \code{FeatureID},
the second is the new \code{FeatureID}.}

\item{fun.aggregate}{Should the data be aggregated before casting? If the formula doesn't identify single observation for each cell, then aggregation defaults to \code{length} with a message.

  \bold{NEW}: it is possible to provide a list of functions to \code{fun.aggregate} argument. See \code{examples}.}

\item{...}{Further arguments passed on to \code{\link[data.table]{dcast}}}
}
\value{
\code{\link{MDT}} object with new \code{FeatureID}s.
All annotations are lost.
}
\description{
Aggregate \code{FeatureID}s into Higher Level IDs.
}
\details{
Melts \code{mtable} using \code{\link[data.table]{melt.data.table}} and
then recasts new \code{mtable} using \code{\link[data.table]{dcast}}.
}

