% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Documentation.R, R/methods-MDT-annotate.R, R/methods-MDT.R
\docType{methods}
\name{annotateStat}
\alias{annotateBioconductor}
\alias{annotateBioconductor,MDT-method}
\alias{annotateCoding}
\alias{annotateCoding,MDT-method}
\alias{annotateGOSim}
\alias{annotateLocation}
\alias{annotateLocation,MDT-method}
\alias{annotateMAF}
\alias{annotateMAF,MDT-method}
\alias{annotateMDT}
\alias{annotateRSID}
\alias{annotateRSID,MDT-method}
\alias{annotateStat}
\alias{annotateStat,MDT-method}
\alias{merge,MDT,data.frame-method}
\alias{merge-MDT-data.frame-method}
\title{Annotate MDT Objects}
\usage{
annotateStat(x, function_list = list(FISHER = function(x, y) {    
  stats::fisher.test(table(factor(x, c(0L, 1L, 2L)), factor(y, c(0L,
  1L))))$p.value }, HW = function(x, y) {    
  HardyWeinberg::HWExact(table(factor(x, c(0L, 1L, 2L))), verbose = FALSE)$pval
  }), fill = NA, verbose = TRUE)

annotateLocation(x, txdb, region = VariantAnnotation::AllVariants(),
  columns = c("LOCATION", "ENTREZID"), verbose = TRUE)

annotateBioconductor(x, annotations_db, keys_colname, columns,
  keytype = keys_colname, verbose = TRUE)

annotateCoding(x, seqSource, txdb, columns = c("CONSEQUENCE"),
  verbose = TRUE)

annotateMAF(x)

annotateRSID(x, snp_loc, verbose = TRUE)

\S4method{annotateMAF}{MDT}(x)

\S4method{annotateStat}{MDT}(x, function_list = list(FISHER = function(x, y) {
      stats::fisher.test(table(factor(x, c(0L, 1L, 2L)), factor(y, c(0L,
  1L))))$p.value }, HW = function(x, y) {    
  HardyWeinberg::HWExact(table(factor(x, c(0L, 1L, 2L))), verbose = FALSE)$pval
  }), fill = NA, verbose = TRUE)

\S4method{annotateLocation}{MDT}(x, txdb,
  region = VariantAnnotation::AllVariants(), columns = c("LOCATION",
  "ENTREZID"), verbose = TRUE)

\S4method{annotateBioconductor}{MDT}(x, annotations_db, keys_colname, columns,
  keytype = keys_colname, verbose = TRUE)

\S4method{annotateCoding}{MDT}(x, seqSource, txdb, columns = c("CONSEQUENCE"),
  verbose = TRUE)

\S4method{annotateRSID}{MDT}(x, snp_loc, verbose = TRUE)

\S4method{merge}{MDT,data.frame}(x, y, by = "FeatureID", all = FALSE,
  all.x = TRUE, all.y = FALSE, sort = TRUE, allow.cartesian = TRUE)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{function_list}{\code{list} of \code{function}s that take two arguments,
a dropped column of \code{mtable(mdt)} and \code{response(mdt)}, and returns
one \code{numeric} value.
\code{list} \code{names} are used as \code{colnames} in
\code{annotations(mdt)}.}

\item{fill}{Value to replace missing values in \code{mtable}.}

\item{verbose}{\code{logical} specifying if function should be run in verbose mode.}

\item{txdb}{\code{\link[GenomicFeatures]{TxDb}} or
\code{\link[GenomicRanges]{GRangesList}} object that
serves as the annotation.
GFF files can be converted to \code{\link[GenomicFeatures]{TxDb}} objects
with \code{makeTxDbFromGF} in the \code{GenomicFeatures} package.}

\item{region}{An instance of one of the 8 VariantType classes :
    \code{CodingVariants}, \code{IntronVariants}, \code{FiveUTRVariants}, 
    \code{ThreeUTRVariants}, \code{IntergenicVariants}, 
    \code{SpliceSiteVariants}, \code{PromoterVariants}, \code{AllVariants}.
    All objects can be instantiated with no arguments, e.g., CodingVariants()
    will create an object of \code{CodingVariants}. 

    \code{AllVariants}, \code{PromoterVariants} and \code{IntergenicVariants}
    have \code{upstream} and \code{downstream} arguments. For
    \code{PromoterVariants} and \code{IntergenicVariants} these are single
    integer values >= 0. For \code{AllVariants} these are integer vectors
    of length 2 named \sQuote{promoter} and \sQuote{intergenic}. See
    ?\code{upstream} for more details.

    When using \code{AllVariants}, a range in \code{query} may fall in
    multiple regions (e.g., 'intergenic' and 'promoter'). In this case
    the result will have a row for each match. All data in the
    row will be equivalent except the LOCATION column. 
  }

\item{columns}{Column names to be appended to \code{annotations(mdt)}.
For \code{annotateLocation}, can me any column returned by
\code{\link[VariantAnnotation]{locateVariants}}.
For \code{annotateCoding}, can be any column returned by
\code{\link[VariantAnnotation]{predictCoding}}.}

\item{annotations_db}{\code{\link[AnnotationDbi]{AnnotationDb}} object.
Can also be \code{\link[VariantAnnotation]{PolyPhenDb}} or
\code{\link[VariantAnnotation]{SIFTDb}}.}

\item{keys_colname}{Name of column to be used as \code{key} argument in
\code{\link[AnnotationDbi]{select}}.}

\item{keytype}{the keytype that matches the keys used.  For the
    \code{select} methods, this is used to indicate the kind of ID being used
    with the keys argument. For the \code{keys} method this is used to
    indicate which kind of keys are desired from \code{keys}
  }

\item{seqSource}{\code{\link[BSgenome]{BSgenome-class}} instance or
a FaFile to be used for sequence extraction.}

\item{snp_loc}{\code{\link[BSgenome]{SNPlocs}} object that contains
known SNP locations for a given organism.}

\item{y}{\code{data.frame} with a column named \code{FeatureID}.}

\item{by}{
    A vector of shared column names in \code{x} and \code{y} to merge on.
    This defaults to the shared key columns between the two tables.
    If \code{y} has no key columns, this defaults to the key of \code{x}.
  }

\item{all}{
    logical; \code{all = TRUE} is shorthand to save setting both \code{all.x = TRUE} and
    \code{all.y = TRUE}.
  }

\item{all.x}{
    logical; if \code{TRUE}, then extra rows will be added to the
    output, one for each row in \code{x} that has no matching row in
    \code{y}.  These rows will have 'NA's in those columns that are
    usually filled with values from \code{y}.  The default is \code{FALSE},
    so that only rows with data from both \code{x} and \code{y} are
    included in the output.
  }

\item{all.y}{
    logical; analogous to \code{all.x} above.
  }

\item{sort}{
    logical. If \code{TRUE} (default), the merged \code{data.table} is sorted 
    by setting the key to the \code{by / by.x} columns. If \code{FALSE}, the 
    result is not sorted.
  }

\item{allow.cartesian}{
    See \code{allow.cartesian} in \code{\link{[.data.table}}.
  }

\item{name}{Name of appended column to \code{annotations(mdt)}.}

\item{target_go}{Target GO list.}
}
\description{
Appends annotations for each
\code{FeatureID} to \code{annotations(mdt)}.
}
\section{merge}{

Merges a \code{data.frame} to annotations using \code{FeatureID}s.
}

\section{annotateStat}{

Estimates association between \code{mtable(mdt)} feature
and \code{response(mdt)}. By default, appends Fisher's exact test p-value and
Hardy-Weinberg equilibrium exact test p-values.
Results can be visualized using \code{\link{manhattanPlot}}.
}

\section{annotateMAF}{

Estimates minor allele frequency.
}

\section{annotateLocation}{

Annotates genomic positions by locating variants
with respect to gene function using
\code{\link[VariantAnnotation]{locateVariants}}.
Please note that \code{GENEID} is renamed as \code{ENTREZID}.
}

\section{annotateBioconductor}{

Annotates variants with Bioconductor
\code{\link[AnnotationDbi]{AnnotationDb}} objects using
\code{\link[AnnotationDbi]{select}}.
}

\section{annotateGOSim}{


Estimates GO semantic similarity scores between a gene vector
and a target GO terms vector.

Uses \code{\link[GOSemSim]{mgoSim}} to estimate GO scores.
\code{annotation(mdt)} must contain \code{GO} and \code{ONTOLOGY} columns
before running function. Please call \code{annotateBioconductor} to do so.
}

\section{annotateRSID}{

Annotates dbSNP RS identification numbers
using \code{\link[GenomicRanges]{findOverlaps}}.
}

\section{annotateCoding}{

Predict coding changes (e.g. synonymous, frameshift) for variants by using
\code{\link[VariantAnnotation]{predictCoding}}.
}
\seealso{
\code{\link[AnnotationDbi]{AnnotationDb}}
\code{\link[VariantAnnotation]{locateVariants}}
\code{\link[GenomicFeatures]{TxDb}}
\code{\link[VariantAnnotation]{predictCoding}}
\code{\link[BSgenome]{BSgenome}}
\code{\link[VariantAnnotation]{PolyPhenDb}}
\code{\link[VariantAnnotation]{SIFTDb}}
\code{\link[BSgenome]{SNPlocs}}
\code{\link[GOSemSim]{mgoSim}}
}

