% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{asMatrixMDT}
\alias{asMatrixMDT}
\alias{asMatrixMDT,MDT-method}
\title{Convert mtable data.table to matrix}
\usage{
asMatrixMDT(x, fill = NA, fun.aggregate = NULL, ...)

\S4method{asMatrixMDT}{MDT}(x, fill = NA, fun.aggregate = NULL, ...)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{fill}{Value to fill missing cells with. If \code{fun.aggregate} is present, takes the value by applying the function on 0-length vector.}

\item{fun.aggregate}{Should the data be aggregated before casting? If the formula doesn't identify single observation for each cell, then aggregation defaults to \code{length} with a message.

  \bold{NEW}: it is possible to provide a list of functions to \code{fun.aggregate} argument. See \code{examples}.}

\item{...}{Any other arguments that maybe passed to the aggregating function.}
}
\value{
Mutation \code{matrix}
}
\description{
Convert mtable \code{\link[data.table]{data.table}} to \code{matrix}
}

