% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtilities.R
\name{chrAsCharacter}
\alias{chrAsCharacter}
\title{Chromsome integer values as character}
\usage{
chrAsCharacter(x)
}
\arguments{
\item{x}{\code{integer} \code{vector} of chromosomes.}
}
\value{
\code{character} \code{vector} of chromosomes.
}
\description{
Chromsome \code{integer} values as \code{character}
}
\details{
Chromosome codes follow MAP files codeing as in PLINK.
Autosomes are coded 1 through 22.
Others use the following code:
\itemize{
\item  X chromosome: X or 23
\item  Y chromosome: Y or 24
\item  Pseudo-autosomal region of X: XY or 25
\item  Mitochondrial: MT or 26
}
}

