% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{fillMDT}
\alias{fillMDT}
\alias{fillMDT,MDT-method}
\title{Fill missing values for MDT objects.}
\usage{
fillMDT(x, fill)

\S4method{fillMDT}{MDT}(x, fill)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{fill}{Value to replace missing values.}
}
\value{
Filled \code{\link{MDT}} object.
}
\description{
Fill in missing values in \code{mtable} slot for \code{\link{MDT}} objects.
Identifies missing combinations of \code{SampleID} and \code{FeatureID} and
gives it value specified in \code{fill} argument. Maybe necessary after
applying \code{\link{filterMDT}}.
}

