% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{filterMDT}
\alias{filterMDT}
\alias{filterMDT,MDT-method}
\title{Filter MDT Objects}
\usage{
filterMDT(x, condition, with = c("annotations", "phenotype", "info"))

\S4method{filterMDT}{MDT}(x, condition, with = c("annotations", "phenotype",
  "info"))
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{condition}{\code{logical} expression indicating elements.
Missing values are taken as \code{FALSE}.}

\item{with}{\code{character(1)} specifying which annotation
\code{\link[data.table]{data.table}} should be used for filtering.}
}
\value{
Filtered \code{\link{MDT}} object.
}
\description{
Filter out samples or features on the basis of the entries in annotations,
phenotype or info.
}

