% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{importPhenotype}
\alias{importPhenotype}
\alias{importPhenotype,MDT-method}
\title{Adds phenotype data to MDT object}
\usage{
importPhenotype(x, file, response_type = "factor", ...)

\S4method{importPhenotype}{MDT}(x, file, response_type = "factor", ...)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{file}{Filename containing phenotypic information.
Must be tab separated, and contain a column named SampleID corresponding
to Sample names, one column named RESPONSE corresponding to the
response variable, and SEX and AGE columns.
Warning: the function is CAPS sensitive.}

\item{response_type}{Class of RESPONSE. Choose between: \code{'numeric'},
\code{'integer'}, \code{'character'} or \code{'factor'}.}

\item{...}{Further arguments to pass on to \code{\link[data.table]{fread}}.}
}
\value{
Updates \code{\link{MDT}} object with phenotype data.
}
\description{
Adds phenotype data to \code{\link{MDT}} object.
}
\details{
Uses \code{\link[data.table]{fread}} to read file.
}
\seealso{
\code{\link[data.table]{fread}}
}

