% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R, R/methods-MLGWAS.R
\docType{methods}
\name{manhattanPlot}
\alias{manhattanPlot}
\alias{manhattanPlot,MDT,ANY-method}
\alias{manhattanPlot,MDT,MLGWAS-method}
\title{Manhattan Plot for MDT and MLGWAS objects.}
\usage{
manhattanPlot(x, value, group = "CHROM", pos = "POS", f = function(value)
  -log10(value), col = c("blue", "orange"))

\S4method{manhattanPlot}{MDT,ANY}(x, value, group = "CHROM", pos = "POS",
  f = function(value) -log10(value), col = c("blue", "orange"))

\S4method{manhattanPlot}{MDT,MLGWAS}(x, value, group = "CHROM", pos = "POS",
  f = function(value) -log10(value), col = c("blue", "orange"))
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{value}{\code{character} specifying which column of \code{annotations(x)}
should be used as values. For example, call \code{annotateStat} and then
use Fisher's exact test p-values. Can also be \code{\link{MLGWAS}}; plots
feature importance.}

\item{group}{\code{character} specifying which column of \code{annotations(x)}
specifying grouping of 'x' axis such as chromosome number or
\code{\link{MLGWAS}} object.}

\item{pos}{\code{character} specifying which column of \code{annotations(x)}
specifying position on 'x' axis.}

\item{f}{Some function that modifies a vector of integers.
For example, \code{-log10} for p-values.}

\item{col}{Plot colors.}
}
\value{
A \code{\link{ggplot}} Manhattan plot.
}
\description{
Plot a Manhattan Plots for \code{\link{MDT}} and \code{\link{MLGWAS}}
objects. Can use any \code{numeric}
value in \code{annotations(x)} as specified by \code{value} argument.
Can also use feature importance as found in \code{featimp(mlgwas)} if
\code{value} is an \code{\link{MLGWAS}} object.
}

