% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MDT-class}
\alias{MDT}
\alias{MDT-class}
\alias{[,MDT,character,character,ANY-method}
\alias{[,MDT,character,missing,ANY-method}
\alias{[,MDT,missing,character,ANY-method}
\alias{[,MDT-method}
\alias{annotations}
\alias{annotations,MDT-method}
\alias{annotations<-}
\alias{annotations<-,MDT-method}
\alias{dim,MDT-method}
\alias{features}
\alias{features,MDT-method}
\alias{features<-}
\alias{features<-,MDT-method}
\alias{info}
\alias{info,MDT-method}
\alias{info<-}
\alias{info<-,MDT-method}
\alias{length,MDT-method}
\alias{merge,MDT-method}
\alias{mtable}
\alias{mtable,MDT-method}
\alias{mtable<-}
\alias{mtable<-,MDT-method}
\alias{ncol,MDT-method}
\alias{nrow,MDT-method}
\alias{phenotype}
\alias{phenotype,MDT-method}
\alias{phenotype<-}
\alias{phenotype<-,MDT-method}
\alias{response}
\alias{response,MDT-method}
\alias{response<-}
\alias{response<-,MDT-method}
\alias{samples}
\alias{samples,MDT-method}
\alias{samples<-}
\alias{samples<-,MDT-method}
\alias{show,MDT-method}
\title{Mutation Data Table}
\arguments{
\item{x}{\code{MDT} object.}

\item{value}{\code{character} vector.}
}
\description{
The MDT class can contain information about sequence variation, annotations
and phenotype for a collection of samples. It is meant to simplify
machine learning in genome-wide association studies.
}
\section{Slots}{

\describe{
\item{\code{mtable}}{Variation matrix.
Must contain the following columns:
\code{FeatureID},
\code{SampleID},
\code{VALUE}.}

\item{\code{annotations}}{Feature annotations.
Must contain the following columns:
\code{FeatureID}.}

\item{\code{phenotype}}{Sample annotations.
Must contain the following columns:
\code{SampleID},
\code{RESPONSE} (variable to be predicted),
\code{SEX},
\code{AGE}.}

\item{\code{info}}{Feature and Sample annotations. Annotations that are dependant on
both sample and feature such as quality measures or observed alleles.
Must contain the following columns:
\code{FeatureID},
\code{SampleID}.}

\item{\code{features}}{\code{character} \code{vector} of unique \code{FeatureID}s.
Each combination of chromosome and position makes one unique \code{FeatureID}.}

\item{\code{samples}}{\code{character} \code{vector} of unique \code{SampleID}s.}
}}
\section{Constructors}{


\code{MDT} contructs a MDT object. Only the mtable argument is required.

\code{\link{vcfsToMDT}} converts a list of VCF objects to MDT without
any phenotypic data, that is, with an empty \code{phenotype} slot.

\code{\link{importPhenotype}} adds phenotypic data to MDT object.

\code{\link{aggregateMDT}} aggregate FeatureIDs Into Higher Level IDs.
}

\section{Accessors}{


\code{mtable(x)}, \code{mtable(x) <- value} gets or sets \code{mtable}.

\code{annotations(x)}, \code{annotations(x) <- value} gets or sets
\code{annotations}.

\code{phenotype(x)}, \code{phenotype(x) <- value} gets or sets
\code{phenotype}.

\code{info(x)}, \code{info(x) <- value} gets or sets \code{info}.

\code{response(x)}, \code{response(x) <- value} gets or sets \code{Response}
column in \code{phenotype} slot.

\code{features(x)}, \code{features(x) <- value} gets or sets
\code{FeatureID}s.

\code{samples(x)}, \code{samples(x) <- value} gets or sets \code{SampleID}s.

\code{asMatrixMDT(x, miss)} converts \code{mtable} tp a
\code{\link[base]{matrix}}
with \code{SampleID} as rows and
\code{FeatureID}s as columns.
}

\section{Subset}{


\code{x[SampleIDs, FeatureIDs]} subsets \code{MDT} objects using \code{SampleID}s
and \code{FeatureID}s. Must be \code{character}.

\code{\link{filterMDT}} filter samples and features in \code{MDT}
objects using annotations in \code{annotations}, \code{phenotype}
and \code{info}.
}

\section{Plots}{


\code{\link{plotMDT}} allows to generate various plots such as PCA, heatmaps,
histograms.

\code{\link{manhattanPlot}} creates a Manhattan plot.
}

\section{Add Annotations}{


See \code{\link{annotateMDT}} to add feature annotations to annotations(x).

See \code{\link{importPhenotype}} to add sample annotations to phenotype(x).

Use \code{\link{selectMDT}} to retrieve annotations from annotations(x),
phenotyope(x) and info(x).
}

\section{Train Classifier}{


\code{\link{trainClassifier}} trains a classifier, using an
\code{MDT} object, that that predicts phenotypic \code{response(x)}
using \code{mtable(x)} matrix. Creates a \code{\link{MLGWAS}} object.
}
\seealso{
\code{\link{MLGWAS}}
\code{\link[data.table]{data.table}}
}

