% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{mdtToGRanges}
\alias{mdtToGRanges}
\alias{mdtToGRanges,MDT-method}
\title{Convert MDT to Genomic Ranges}
\usage{
mdtToGRanges(x)

\S4method{mdtToGRanges}{MDT}(x)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}
}
\value{
\code{\link[GenomicRanges]{GRanges}} object.
}
\description{
Converts \code{\link{MDT}} to \code{\link[GenomicRanges]{GRanges}}
}

