% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MLGWAS-class}
\alias{MLGWAS}
\alias{MLGWAS-class}
\alias{[,MLGWAS,character,ANY,ANY-method}
\alias{[,MLGWAS,logical,ANY,ANY-method}
\alias{[,MLGWAS,numeric,ANY,ANY-method}
\alias{[,MLGWAS-method}
\alias{bindML}
\alias{bindML,MLGWAS-method}
\alias{classifierNames}
\alias{classifierNames,MLGWAS-method}
\alias{classifierNames<-}
\alias{classifierNames<-,MLGWAS-method}
\alias{classifiers}
\alias{classifiers,MLGWAS-method}
\alias{classifiers<-}
\alias{classifiers<-,MLGWAS-method}
\alias{dim,MLGWAS-method}
\alias{errors}
\alias{errors,MLGWAS-method}
\alias{errors<-}
\alias{errors<-,MLGWAS-method}
\alias{featimp}
\alias{featimp,MLGWAS-method}
\alias{featimp<-}
\alias{featimp<-,MLGWAS-method}
\alias{features,MLGWAS-method}
\alias{features<-,MLGWAS-method}
\alias{length,MLGWAS-method}
\alias{ncol,MLGWAS-method}
\alias{nrow,MLGWAS-method}
\alias{partitionNames}
\alias{partitionNames,MLGWAS-method}
\alias{partitionNames<-}
\alias{partitionNames<-,MLGWAS-method}
\alias{partitions}
\alias{partitions,MLGWAS-method}
\alias{partitions<-}
\alias{partitions<-,MLGWAS-method}
\alias{performances}
\alias{performances,MLGWAS-method}
\alias{performances<-}
\alias{performances<-,MLGWAS-method}
\alias{predictions}
\alias{predictions,MLGWAS-method}
\alias{predictions<-}
\alias{predictions<-,MLGWAS-method}
\alias{samples,MLGWAS-method}
\alias{samples<-,MLGWAS-method}
\alias{show,MLGWAS-method}
\title{Machine Learning for Genome Wide Association Studies (MLGWAS)}
\arguments{
\item{x}{\code{MLGWAS} object.}
}
\description{
The MLGWAS contains one or more classifiers with information about their
global performance, local errors and feature importance.
It is meant to simplify machine learning in genome-wide association studies.
}
\section{Slots}{

\describe{
\item{\code{performances}}{Global performance measures of classifiers.
First two columns are \code{ClassifierID}s and \code{PartitionID}s.
Other columns are performance measures.
For categorical responses, these are:
sensitivity, specificity,
positive predictive value (PPV), negative predictive value (NPV),
ROC plot AUC and F1 score.
For continuous respones, these are:
root-mean-square deviation (RMSD) and coefficient of determination (r2).}

\item{\code{featimp}}{Variable importance of \code{FeatureID}
as estimated by \code{\link[caret]{varImp}}.
Columns are \code{ClassifierID}, \code{PartitionID},
\code{FeatureID} and \code{Importance}.}

\item{\code{errors}}{Errors per partition and per sample.
Columns are \code{ClassifierID}, \code{PartitionID},
\code{SampleID} and \code{Error}.
For categorical variables, values are integers: 1 correspond to a mistake,
0 to a success.
For continuous, values the square of the difference between the prediction
and the actual response.
In both cases, NA means that this sample was not part of the test set.}

\item{\code{classifiers}}{Classifiers.
Coded as a \code{list} of \code{list}s of \code{\link[caret]{train}}
objects.
First level corresponds to \code{ClassifierID}
and second level to \code{PartitionID}.}

\item{\code{partitions}}{Train and test set partitions
Coded as a \code{list} of \code{list}s.
First level corresponds to \code{ClassifierID}s and second level to
\code{PartitionID}s.}

\item{\code{predictions}}{Predictions of \code{response(mdt)} for test set.
Coded as a \code{list} of \code{list}s.
First level corresponds to \code{ClassifierID}s and second level to
\code{PartitionID}s.}

\item{\code{call}}{List of calls. Names of list are \code{ClassifierIDs}}
}}
\section{Constructors}{


\code{\link{trainClassifier}} trains a classifier,
using an \code{\link{MDT}} object,
that that predicts \code{response(mdt)} affection using \code{mtable(mdt)} or
\code{phenotype(mdt)} or both.

\code{\link{trainClassifiers}} allows to call \code{\link{trainClassifier}}
using different combination of parameters specified as a \code{data.frame}.
}

\section{Identifiers}{


Four identifiers are present in this class:
\itemize{
\item \code{ClassifierID} as specified by the \code{id} argument
in \code{\link{trainClassifier}}.
\item \code{PartitionID} as specified by the \code{names} of the
output of the \code{partition} argument in \code{\link{trainClassifier}}.
\item \code{FeatureID} as specified by the \code{FeatureID}s
of the \code{MDT} object.
\item \code{SampleID} as specified by the \code{Sample}s
of the \code{MDT} object.
}
}

\section{Subset and Combine}{


\code{x[ClassifierIDs]} subsets \code{MLGWAS} objects. Must be a \code{character}.

\code{bindML(...)} combines \code{MLGWAS} objects.
}

\section{Summaries}{


The following functions summarize results throughout partitions:

\code{\link{summaryPerf}} summarizes performance per classifier.

\code{\link{summaryFeatimp}} summarizes feature importance per classifier.

\code{\link{summaryErrors}} summarizes errors per classifier and per sample.
}

\section{Plots}{


\code{\link{plotPerf}} plots performance as boxplots.

\code{\link{plotFeatimp}} plots feature importance as boxplots.

\code{\link{plotErrors}} plots sample error as boxplots.

\code{\link{manhattanPlot}} creates a Manhattan plot using the variable
importance of features as estimated by \code{\link[caret]{varImp}} instead
of -log10 pvalues.
}

\section{Accessors}{


\code{performances(x)} and \code{performances(x) <- value} gets or sets
\code{performances}.

\code{featimp(x)} and \code{featimp(x) <- value} gets or sets
\code{featimp}.

\code{errors(x)} and \code{errors(x) <- value} gets or sets
\code{errors}.

\code{classifiers(x)} and \code{classifiers(x) <- value} gets or sets
\code{classifiers}.

\code{partitions(x)} and \code{partitions(x) <- value} gets or sets
\code{partitions}.

\code{predictions(x)} and \code{predictions(x) <- value} gets or sets
\code{predictions}.

\code{classifierNames(x)} and \code{classifierNames(x) <- value} gets or sets
\code{ClassifiersID}s.

\code{partitionNames(x)} and \code{partitionNames(x) <- value}
get or sets \code{PartitionID}s.

\code{features(x)} and \code{features(x) <- value} get or sets
\code{FeatureID}.

\code{samples(x)} and \code{samples(x) <- value} get or sets
\code{SampleID}s
}
\seealso{
\code{\link{trainClassifier}}
\code{\link{trainClassifiers}}
\code{\link{MDT}}
\code{\link[caret]{train}}
\code{\link[data.table]{data.table}}
}

