% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{plotMDT}
\alias{plotMDT}
\alias{plotMDT,MDT-method}
\title{Plot MDT objects}
\usage{
plotMDT(x, type = c("scatter", "heatmap"), preprocess = c("scale", "center",
  "pca"), space = c("samples", "features"), dims = c(1L, 2L), fill = NA,
  ...)

\S4method{plotMDT}{MDT}(x, type = c("scatter", "heatmap"),
  preprocess = c("scale", "center", "pca"), space = c("samples",
  "features"), dims = c(1L, 2L), fill = NA, ...)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{type}{Type of plot
\itemize{
\item{scatter: }{Scatter plot using \code{\link[ggplot2]{geom_point}}}
\item{heatmap: }{Heatmap using \code{\link[gplots]{heatmap.2}}}
}}

\item{preprocess}{Preprocessing transformation. Choose from 'scale',
'center' and 'pca'.}

\item{space}{For \code{y='scatter'}, specifies
if should plot feature space or sample space.}

\item{dims}{For \code{y='scatter'}, specifies which features or samples
should be plotted. Can be integers specifying column index or
characters.}

\item{fill}{Value to replace missing values in \code{mtable}.}

\item{...}{Further arguments to be passed on to
\code{\link[gplots]{heatmap.2}} or \code{\link[ggplot2]{geom_point}}}
}
\description{
Various plots for \code{\link{MDT}} objects.
}

