% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtilities.R
\name{readVcfDir}
\alias{readVcfDir}
\title{Reads directory of VCF files}
\usage{
readVcfDir(dir, param_function = function(fn)
  VariantAnnotation::ScanVcfParam(), grep_out, n_files, verbose = TRUE, ...)
}
\arguments{
\item{dir}{Directory containing VCF files.}

\item{param_function}{Function to be evaluated and as param argument.
Only takes as argument filename and must return a
\code{\link[VariantAnnotation]{ScanVcfParam}} object.}

\item{grep_out}{Regular expression for file names to be filter out.}

\item{n_files}{If specified, will only import the first
\code{n_files} in the directory.}

\item{verbose}{\code{logical} specifying if function should be run in verbose mode.}

\item{...}{Further arguments to pass on to
\code{\link[VariantAnnotation]{readVcf}}}
}
\value{
A list of \code{\link[VariantAnnotation]{VCF}} objects.
}
\description{
Real all Variant Call Format (VCF) files contained in a
directory using the VariantAnnotation Bioconductor package.
}
\details{
Uses \code{\link{list.files}} to identify files ending with ".vcf"'.
}
\seealso{
\code{\link[VariantAnnotation]{readVcf}}
}

