% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{selectMDT}
\alias{selectMDT}
\alias{selectMDT,MDT-method}
\title{Retrieve annotations from MDT objects.}
\usage{
selectMDT(x, keys, columns, keytype = "annotations", na.rm = TRUE)

\S4method{selectMDT}{MDT}(x, keys, columns, keytype = "annotations",
  na.rm = TRUE)
}
\arguments{
\item{x}{\code{\link{MDT}} object.}

\item{keys}{\code{character} or \code{\link{data.table}} with SampleID
and/or FeatureID key values to select records.
Can only be a \code{character} when \code{keytype} is not \code{info}.}

\item{columns}{The columns to be retrieved from the annotation data.table.}

\item{keytype}{\code{character} specifying which annotation data.table
should be used. Choose from \code{"annotations"}, \code{"phenotype"} or
\code{"info"}.}

\item{na.rm}{Should NA values be removed from results.}
}
\value{
\code{\link[data.table]{data.table}} with relevant key and annotations.
}
\description{
Retrieve annotations from annotations(x), phenotyope(x) and info(x) in MDT objects.
}

