% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Documentation.R, R/methods-MLGWAS.R
\docType{methods}
\name{summaryPerf}
\alias{summaryErrors}
\alias{summaryErrors,MLGWAS-method}
\alias{summaryFeatimp}
\alias{summaryFeatimp,MLGWAS-method}
\alias{summaryMLGWAS}
\alias{summaryPerf}
\alias{summaryPerf,MLGWAS-method}
\title{Summary of MLGWAS Performance, Feature Importance and Errors}
\usage{
summaryPerf(x, summarize = function(x) c(mean = mean(x, na.rm = TRUE), sd =
  stats::sd(x, na.rm = TRUE), len = length(stats::na.omit(x))))

summaryFeatimp(x, summarize = function(x) c(mean = mean(x, na.rm = TRUE), sd =
  stats::sd(x, na.rm = TRUE), len = length(stats::na.omit(x))))

summaryErrors(x, summarize = function(x) c(mean = mean(x, na.rm = TRUE), sd =
  stats::sd(x, na.rm = TRUE), len = length(stats::na.omit(x))))

\S4method{summaryPerf}{MLGWAS}(x, summarize = function(x) c(mean = mean(x,
  na.rm = TRUE), sd = stats::sd(x, na.rm = TRUE), len =
  length(stats::na.omit(x))))

\S4method{summaryFeatimp}{MLGWAS}(x, summarize = function(x) c(mean = mean(x,
  na.rm = TRUE), sd = stats::sd(x, na.rm = TRUE), len =
  length(stats::na.omit(x))))

\S4method{summaryErrors}{MLGWAS}(x, summarize = function(x) c(mean = mean(x,
  na.rm = TRUE), sd = stats::sd(x, na.rm = TRUE), len =
  length(stats::na.omit(x))))
}
\arguments{
\item{x}{\code{\link{MLGWAS}} object.}

\item{summarize}{\code{list} of \code{function}s to summarize performance.
Each function must take a vector of \code{numeric}s
and return one \code{numeric} value.}
}
\value{
Summarized \code{\link{data.table}}
}
\description{
Summarizes performance, feature importance or errors of
\code{\link{MLGWAS}} objects throughout partitions.
}

