% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MLGWAS.R
\docType{methods}
\name{trainClassifier}
\alias{trainClassifier}
\alias{trainClassifier,MDT-method}
\title{Train Classifier for MDT}
\usage{
trainClassifier(mdt, id, mtable_vars = "all", phen_vars = "none",
  fill = NA, partitions = caret::createMultiFolds(y = response(mdt), k =
  10L, times = 1L), preprocess = function(x, y) function(x) x,
  method = "rf", permute = FALSE, validation = TRUE, parallel = FALSE,
  verbose = TRUE, .export = NULL, ...)

\S4method{trainClassifier}{MDT}(mdt, id, mtable_vars = "all",
  phen_vars = "none", fill = NA, partitions = caret::createMultiFolds(y =
  response(mdt), k = 10L, times = 1L), preprocess = function(x, y) function(x)
  x, method = "rf", permute = FALSE, validation = TRUE,
  parallel = FALSE, verbose = TRUE, .export = NULL, ...)
}
\arguments{
\item{mdt}{\code{\link{MDT}} object.
\code{FeatureID}s and \code{SampleID}s will correspond between the
\code{\link{MDT}} input and the \code{\link{MLGWAS}} output object.}

\item{id}{Identification name of classifier: \code{ClassifierID}. By default,
uses the name of the method appended with '.permuted' if applicable.}

\item{mtable_vars}{Variables in \code{mtable} to be included in matrix.
Choose between \code{'none'}, \code{'all'}
or a subset of values in \code{features(x)}.}

\item{phen_vars}{Variables in \code{phenotype} to be included in matrix.
Choose between \code{'none'}, \code{'all'} or a subset of values in
\code{colnames(phenotype(x))}. \code{SampleID} and \code{Response}
 are excluded.}

\item{fill}{How should missing values in \code{\link{MDT}} be handled.
By default, does nothing: leaves then as \code{NA}.}

\item{partitions}{Some \code{function} that takes a \code{\link{MDT}}
object as input and returns a \code{list} of \code{integer}s specifying
the training set indexes.
See \code{\link[caret]{createDataPartition}} for such functions.
Names of \code{list} will define the \code{PartitionID}.}

\item{preprocess}{Some \code{function} that takes the training matrix,
with features as columns and samples as rows,
and the response (response(mdt)) as input and returns a
\code{function} that modifies a
a matrix with the same number of columns (features).
Can be used for feature selection or preprocessing.
The function will be created internally only using the training partition,
thus allowing to apply the feature selection to the test set without using
any of its information.}

\item{method}{a string specifying which classification or regression model
to use. Possible values are found using \code{names(getModelInfo())}. See
\url{http://topepo.github.io/caret/bytag.html}. A list of functions can also
be passed for a custom model function. See
\url{http://topepo.github.io/caret/custom_models.html} for details.}

\item{permute}{\code{logical} specifying if \code{\link{MDT}} should be
permuted. Distribution of performance scores will correspond to a null
distribution and can be used to test the significance of the a classifier.}

\item{validation}{\code{logical} specifying if test set should be data
samples not in \code{partition}.}

\item{parallel}{\code{logical} specifying if parallel interface should be used.
Depends on \code{\link[foreach]{foreach}}.
See \url{https://cran.r-project.org/web/packages/doParallel/vignettes/gettingstartedParallel.pdf}}

\item{verbose}{\code{logical} specifying if function should be run in verbose mode.}

\item{.export}{character vector of variables to export.
    This can be useful when accessing a variable that isn't defined in the
    current environment.
    The default value in \code{NULL}.}

\item{...}{Further arguments to pass on to \code{\link[caret]{train}}.}
}
\value{
\code{\link{MLGWAS}} object.
}
\description{
Trains a classifier that predicts phenotypic \code{response} for
\code{\link{MDT}} objects.
}
\seealso{
\code{\link{trainClassifiers}}
\code{\link[caret]{train}}
\code{\link[caret]{createDataPartition}}
}

