% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MDT.R
\docType{methods}
\name{vcfsToMDT}
\alias{vcfsToMDT}
\alias{vcfsToMDT,list-method}
\title{Convert a list of VCF objects into an MDT object}
\usage{
vcfsToMDT(vcfs, fill = 0L, verbose = TRUE)

\S4method{vcfsToMDT}{list}(vcfs, fill = 0L, verbose = TRUE)
}
\arguments{
\item{vcfs}{List of \code{\link[VariantAnnotation]{VCF}} objects.}

\item{fill}{Value for entries absent in VCF file. By default, sets to "0/0"
corresponding to homozygote for REF allele.}

\item{verbose}{\code{logical} specifying if function should be run in
verbose mode.}
}
\value{
\code{\link{MDT}} object with \code{phenotype} slot.
}
\description{
Convert a \code{list} of
\code{\link[VariantAnnotation]{VCF}} objects, each representing an individual
patient, into a \code{\link{MDT}} object.
}
\details{
Please note that chromosome X, Y, XY and M/MT are replaced by integers
23, 24, 25 and 26. All other chromosome names are replaced by 99.
Please use \code{\link{importPhenotype}} to fill in \code{phenotype} slot.
}
\seealso{
\code{\link[VariantAnnotation]{VCF}}
\code{\link[VariantAnnotation]{readVcf}}
\code{\link{readVcfDir}}
\code{\link{importPhenotype}}
}

